bat: Move random file creation to common BAT code
By moving createRandomFile from ciao-image to CreateRandomFile in the
common BAT package it can be reused for storage based testing.
